myapp/requirements.txt
asgiref==3.11.1
dj-database-url==3.0.1
Django==5.0.6
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
pillow==12.1.1
psycopg2-binary==2.9.11
PyJWT==2.12.1
sqlparse==0.5.5
tzdata==2025.3
gunicorn==21.2.0
